Migrate to new Data_oM project, move the old DataStructure_oM objects into a `Structures` namespace
