+++ Data_oM/Data_oM.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{9ABA7EFB-1B0C-422D-B9B9-191E8AFB7FA8}</ProjectGuid>
+    <ProjectGuid>{239C5F01-F82E-4CCE-8598-5134B7756247}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>BH.oM.DataManipulation</RootNamespace>
-    <AssemblyName>DataManipulation_oM</AssemblyName>
+    <RootNamespace>BH.oM.DataStructure</RootNamespace>
+    <AssemblyName>DataStructure_oM</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -41,16 +41,31 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Structures\DBSCANObject.cs" />
+    <Compile Include="Structures\Graph.cs" />
+    <Compile Include="Structures\GraphLink.cs" />
+    <Compile Include="Structures\GraphNode.cs" />
+    <Compile Include="Structures\IDataStructure.cs" />
+    <Compile Include="Structures\DiscretePoint.cs" />
+    <Compile Include="Structures\LocalData.cs" />
+    <Compile Include="Structures\PointMatrix.cs" />
+    <Compile Include="Structures\PriorityQueue.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Requests\BatchRequest.cs" />
     <Compile Include="Requests\CustomRequest.cs" />
     <Compile Include="Requests\FilterRequest.cs" />
     <Compile Include="Requests\IRequest.cs" />
+    <Compile Include="Structures\Tree.cs" />
+    <Compile Include="Structures\VennDiagram.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\BHoM\BHoM.csproj">
       <Project>{94d88927-62a2-48fc-8efe-d139b67a3373}</Project>
       <Name>BHoM</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Geometry_oM\Geometry_oM.csproj">
+      <Project>{5B09A2E5-CBF5-43E2-8D98-B5035391DB7B}</Project>
+      <Name>Geometry_oM</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
