+++ Data_oM/Structures/DiscretePoint.cs
@@ -20,20 +20,54 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System.Collections.Generic;
+using System;
 
-namespace BH.oM.DataStructure
+namespace BH.oM.Data.Structures
 {
-    public class PointMatrix<T> : IDataStructure
+    public class DiscretePoint : IDataStructure, IComparable<DiscretePoint>
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        public double CellSize { get; set; } = 1.0;
+        public int X { get; set; } = 0;
+        public int Y { get; set; } = 0;
+        public int Z { get; set; } = 0;
 
-        public Dictionary<DiscretePoint, List<LocalData<T>>> Data { get; set; } = new Dictionary<DiscretePoint, List<LocalData<T>>>();
-        
+
+        /***************************************************/
+        /**** IComparable Interface                     ****/
+        /***************************************************/
+
+        public int CompareTo(DiscretePoint other)
+        {
+            if (X != other.X)
+                return X.CompareTo(other.X);
+            else if (Y != other.Y)
+                return Y.CompareTo(other.Y);
+            else
+                return Z.CompareTo(other.Z);
+        }
+
+        /***************************************************/
+
+        public override int GetHashCode()
+        {
+            return X.GetHashCode() ^ Y.GetHashCode() ^ Z.GetHashCode();
+        }
+
+        /***************************************************/
+
+        public override bool Equals(object obj)
+        {
+            if (obj is DiscretePoint)
+            {
+                DiscretePoint other = (DiscretePoint)obj;
+                return ((this.X == other.X) && (this.Y == other.Y) && (this.Z == other.Z));
+            }
+            return false;
+        }
+
         /***************************************************/
     }
 }
