+++ Data_oM/Structures/Graph.cs
@@ -20,19 +20,19 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-namespace BH.oM.DataStructure
+using System.Collections.Generic;
+
+namespace BH.oM.Data.Structures
 {
-    public class GraphLink<T> : IDataStructure
+    public class Graph<T> : IDataStructure
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        public double Weight { get; set; } = 1.0;
+        public List<GraphNode<T>> Nodes { get; set; } = new List<GraphNode<T>>();
 
-        public GraphNode<T> StartNode { get; set; } = new GraphNode<T>();
-
-        public GraphNode<T> EndNode { get; set; } = new GraphNode<T>();
+        public List<GraphLink<T>> Links { get; set; } = new List<GraphLink<T>>();
 
 
         /***************************************************/
